4조 불협화음 2/requirements.txt
Flask==2.0.1
google-cloud-vision==2.4.2
opencv-python-headless
google-cloud-aiplatform
google-search-results
flask-cors
google-cloud-pubsub

# main_v9.py -> vertex ai
# main_v10.py -> cloud vision api